First prototype
This initial prototype, built with Kivy – a cross-platform Python interface library: https://kivy.org/ – provides a graph editor along with two text editors. The graph editor enables nodes and edges to be created. Nodes, for now, are no more than a uniquely named Python class that implements a constructor. Each node has a variable number of inputs and outputs. Edges connect the output of one node to the input of another. Once ready, the graph can be assembled into a program: each node class is written to file, along with a set of executions (a recursive walk through the edges starting from a leaf node). The two text editors enable a nodes documentation and implementation to be configured. Finally, this prototype provides operations to export and import graphs, to save the program represented by the graph in its assembled form, and also to execute the program (displaying the result to a status bar).