Prototype editor, interface wrapper, and menu action [note]
When an action is trigger from the menu, the editor works with the client interface to render the effect of that action. In this example, when a new node is requested the menu sends that request on to the editor. The editor asks the client to create a new node and, in turn, the client asks the server to complete that action. The client then passes the resulting information about the new node back to the editor, which then renders that node in the graphical display (for now as nothing more than a red rectangle).