Improve visual design [note]
The menu can now be open and closed by clicking its right-hand side. The colors used to present the editor, menu, nodes, and variables are now based on the Cyanogen Color Palette (https://www.color-hex.com/color-palette/78826). The text used for nodes, variables, and the menu has been configured. And, finally, configuration files offer a means to change most of the above.