Implement export-to and import-from file [note]
I've added two new menu buttons to support importing a graph from JSON and exporting the current graph to JSON. The export button simply asks the server to serialize the current graph as JSON, prompts the user for a filename, and then provides a "download" option that will save obtained JSON string to the specified file. Users can drag-n-drop a correctly formatted JSON file onto the import button, which will provoke the server to create a new graph object from the file's contents. The client is then refreshed to reflect the new graph.