Fix - ensure connectors update after execution [note]
Since execution runs sever-side, some amount of time passes between triggering execution and getting a response. Importantly, in this time, one can move the mouse outside. Consequently, the value for the `editor.lastHovered` field might change between triggering execution and getting the response. To avoid this issue, I now save the value of last hovered to a local variable before invoking the execution. The callback uses that local instead of the editor's field.